+++ azurerm/__init__.py
@@ -7,6 +7,7 @@ from .container import *
 from .computerp import *
 from .cosmosdbrp import *
 from .deployments import *
+from .graphfns import *
 from .insightsrp import *
 from .keyvault import *
 from .networkrp import *